pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "discord-py>=2.5.2",
    "flask>=3.1.1",
    "psycopg2-binary>=2.9.10",
    "requests>=2.32.4",
]
